// src/types/types.ts
export interface SignInUserInterface {
  phone_number: string; // Phone number with a max length of 10 and min length of 10
  password: string; // Password with a max length of 68 and min length of 6
}